requirements.txt
pandas
numpy
python-dotenv
pymysql
scikit-learn